Refactor TextDocument tests into separate classes
